Removed tooltip.exit().remove() from tooltip.js
Verify if this is the best way to fix the missing tooltip issue.
When more than 2 different chart types on the same page and one is a stackedAreaChart the tooltip would stop working.
When mouseover a lineBarChart and then back to mouseover a stackedAreaChart the tooltip div would already be deleted.